# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vetrotech",
        "contact": {
            "email": "P1@zetorzeszow.pl"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/kontrahent": {
            "get": {
                "tags": [
                    "Kontrahent"
                ],
                "summary": "Wyświetl listę kontrahentów",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Kontrahent"
                ],
                "summary": "Dodaj kontrahenta",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "boolean",
                                        "example": "0"
                                    },
                                    "companyName": {
                                        "type": "string",
                                        "example": "Przykładowy nazwa firmy"
                                    },
                                    "lname": {
                                        "type": "string",
                                        "example": "Jan"
                                    },
                                    "fname": {
                                        "type": "string",
                                        "example": "Kowalski"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Rzeszów"
                                    },
                                    "postalCode": {
                                        "type": "string",
                                        "example": "35-005"
                                    },
                                    "street": {
                                        "type": "string",
                                        "example": "aleja Tadeusza Rejtana"
                                    },
                                    "numberHouse": {
                                        "type": "string",
                                        "example": "55"
                                    },
                                    "numberApartment": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "numberPhone": {
                                        "type": "integer",
                                        "example": "17 282 50 00"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "example@zetorzeszow.pl"
                                    },
                                    "comments": {
                                        "type": "string",
                                        "example": "Uwagi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "boolean"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "An \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kontrahent/{id}": {
            "get": {
                "tags": [
                    "Kontrahent"
                ],
                "summary": "Wyświetl kontrahenta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "An \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Kontrahent"
                ],
                "summary": "Edytuj kontrahenta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "boolean",
                                        "example": "0"
                                    },
                                    "companyName": {
                                        "type": "string",
                                        "example": "Przykładowy nazwa firmy"
                                    },
                                    "lname": {
                                        "type": "string",
                                        "example": "Jan"
                                    },
                                    "fname": {
                                        "type": "string",
                                        "example": "Kowalski"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Rzeszów"
                                    },
                                    "postalCode": {
                                        "type": "string",
                                        "example": "35-005"
                                    },
                                    "street": {
                                        "type": "string",
                                        "example": "aleja Tadeusza Rejtana"
                                    },
                                    "numberHouse": {
                                        "type": "string",
                                        "example": "55"
                                    },
                                    "numberApartment": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "numberPhone": {
                                        "type": "integer",
                                        "example": "17 282 50 00"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "example@zetorzeszow.pl"
                                    },
                                    "comments": {
                                        "type": "string",
                                        "example": "Uwagi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "boolean"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Kontrahent"
                ],
                "summary": "Usuń kontrahenta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kontrahent/allContractor": {
            "get": {
                "tags": [
                    "Usunięci kontrahenci"
                ],
                "summary": "Lista wszystkich utworzonych kontrahentów",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kontrahent/deleted": {
            "get": {
                "tags": [
                    "Usunięci kontrahenci"
                ],
                "summary": "Lista usuniętych kontrahentów",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/kontrahent/renew/{id}": {
            "get": {
                "tags": [
                    "Usunięci kontrahenci"
                ],
                "summary": "Przywróć kontrahenta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Error"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}